[REFACTOR] Move Comment File to New Comment Directory For Enhance Management
